backend/requirements.txt
fastapi
uvicorn
openai
gspread_asyncio
google-oauth
python-dotenv
nest_asyncio